Rework of Sync Committee Duty Tables (Dashboard and Live Monitoring)
Optimizer for multiple validators in the same sync committee by not showing each participation as a separate row but instead showing which validators participated or missed in this slot.
Exclude pending sync committee duties (dashboard).
Add the average time until the next proposal to the dashboard.
Minor style and wording changes.